requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
birdy==0.2
elasticsearch==1.3.0
feedparser==5.1.3
itsdangerous==0.24
oauthlib==0.7.2
python-dateutil==2.4.0
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
urllib3==1.10
wsgiref==0.1.2
